requirements.txt
asgiref==3.8.1
Django==5.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environs==11.0.0
marshmallow==3.21.3
packaging==24.1
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
